se creo componente especifico para la creacion de una solicitud, y se ajusto la inclusion de este modulo en solicitud-tercer y evaluar solicitud para eliminar codigo redundante
