+++ src/app/pages/apoyo-alimentario/inscripciones/crear-solicitud/crear-solicitud.component.scss
@@ -1,0 +1,89 @@
+@import "../../apoyo-alimentario.component.scss";
+.label-color-ud {
+    color: #4c1818;
+    font-weight: bold;
+    background-color: transparent;
+    border-color: transparent;
+    margin-right: 2px;
+    margin-left: -12px;
+}
+
+nb-card.ng-star-inserted {
+    padding-bottom: -5%;
+}
+
+.line {
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+    border-top: 1;
+    border-top: 1px solid rgba(0, 0, 0, 0.1);
+}
+
+nb-card.ng-star-inserted hr {
+    overflow: hidden;
+    margin-top: 0;
+    margin-bottom: 0;
+    border-top: 0;
+}
+
+.non-selectable {
+    -moz-user-select: none;
+    -webkit-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+}
+
+.form-group button {
+    align-items: flex-end;
+    align-content: flex-end;
+}
+
+.form-control {
+    display: inline-flex;
+    flex-wrap: wrap;
+    align-items: center;
+    align-content: center;
+    text-align: center;
+}
+
+.form-control input select {
+    width: 83%;
+}
+
+.input-group-prepend {
+    width: 18%;
+}
+
+@media (max-width: 770px) {
+    .input-group-prepend {
+        width: 30%;
+    }
+    .form-control input select {
+        width: 70%;
+    }
+    .small-ud {
+        margin-left: 28%;
+    }
+}
+
+@media (max-width: 575px) {
+    .input-group-prepend,
+    .form-control input select {
+        width: 55%;
+    }
+    .small-ud {
+        margin-left: 1%;
+    }
+}
+
+.item-list {
+    align-items: center;
+    margin-bottom: 10px;
+    &:last-child {
+        margin-bottom: 0;
+    }
+}
+
+nb-accordion {
+    box-shadow: none;
+}
